Add transitions for all node types
The NullNode and BranchNode editors were just a dropdown so they were
removed and just included directly in the Sidebar
