+++ src/components/Sidebar/index.ts
@@ -1,9 +1,8 @@
 export { default } from './Sidebar';
 export { default as SidebarHeader } from './SidebarHeader';
 export { default as SidebarButton } from './SidebarButton';
-export { default as BranchNodeEditor } from './BranchNodeEditor';
 export { default as ActionNodeEditor } from './ActionNodeEditor';
-export { default as NullNodeEditor } from './NullNodeEditor';
 export { default as BranchTransition } from './BranchTransition';
 export { default as OutlinedDiv } from './OutlinedDiv';
 export { default as DeleteModal } from './DeleteModal';
+export { default as TransitionEditor } from './TransitionEditor';
